+++ app/interactors/delete_static_invite.rb
@@ -1,0 +1,10 @@
+# Delete StaticInvite
+class DeleteStaticInvite
+  include Interactor
+
+  # required context
+  # context.static_invite
+  def call
+    context.static_invite.destroy
+  end
+end